+++ views.py
@@ -10,7 +10,6 @@ import flask
 import urllib.parse
 import matplotlib
 import matplotlib.pyplot as plt
-import numba as nb
 import numpy as np
 import qrcode
 import werkzeug
@@ -18,6 +17,8 @@ from spectrum_utils import plot as sup, spectrum as sus
 
 import parsing
 import drawing
+import similarity
+
 from error import UsiError
 from config import USI_SERVER
 
@@ -40,11 +41,6 @@ default_plotting_args = {
 }
 
 blueprint = flask.Blueprint('ui', __name__)
-
-
-SpectrumTuple = collections.namedtuple(
-    'SpectrumTuple', ['precursor_mz', 'precursor_charge', 'mz', 'intensity'])
-
 
 @blueprint.route('/', methods=['GET'])
 def render_homepage():
@@ -168,7 +164,7 @@ def _generate_figure(spectrum: sus.MsmsSpectrum, extension: str,
 
     try:
         import tasks
-        result = tasks.generate_figure.apply_async(args=[spectrum, extension, kwargs], serializer="pickle")
+        result = tasks.task_generate_figure.apply_async(args=[spectrum, extension, kwargs], serializer="pickle")
         return result.get()
     except:
         return drawing.generate_figure(spectrum, extension, **kwargs)
@@ -198,104 +194,13 @@ def _generate_mirror_figure(spectrum_top: sus.MsmsSpectrum,
     io.BytesIO
         Bytes buffer containing the mirror plot.
     """
-    usi1 = spectrum_top.identifier
-    usi2 = spectrum_bottom.identifier
 
-    fig, ax = plt.subplots(figsize=(kwargs['width'], kwargs['height']))
-
-    # Determine cosine similarity and matching peaks.
-    if kwargs['cosine']:
-        # Assign the matching peak annotations.
-        similarity, peak_matches = _cosine(
-            spectrum_top, spectrum_bottom, kwargs['fragment_mz_tolerance'],
-            kwargs['cosine'] == 'shifted')
-        peak_matches = zip(*peak_matches)
-    else:
-        similarity = 0
-        # Make sure that top and bottom spectra are colored..
-        peak_matches = [np.arange(len(spectrum_top.annotation)),
-                        np.arange(len(spectrum_bottom.annotation))]
-
-    if spectrum_top.peptide is None:
-        # Initialize the annotations as unmatched.
-        for annotation in spectrum_top.annotation:
-            if annotation is not None:
-                annotation.ion_type = 'unmatched'
-        for annotation in spectrum_bottom.annotation:
-            if annotation is not None:
-                annotation.ion_type = 'unmatched'
-
-        for peak_idx, spectrum, label in zip(peak_matches,
-                                             [spectrum_top, spectrum_bottom],
-                                             ['top', 'bottom']):
-            for i in peak_idx:
-                if spectrum.annotation[i] is None:
-                    spectrum.annotation[i] = sus.FragmentAnnotation(
-                        0, spectrum.mz[i], '')
-                spectrum.annotation[i].ion_type = label
-
-        # Colors for mirror plot peaks (subject to change).
-        sup.colors['top'] = '#212121'
-        sup.colors['bottom'] = '#388E3C'
-        sup.colors['unmatched'] = 'darkgray'
-        sup.colors[None] = 'darkgray'
-
-        sup.mirror(spectrum_top, spectrum_bottom,
-                   {'annotate_ions': kwargs['annotate_peaks'],
-                    'annot_kws': {'rotation': kwargs['annotation_rotation'],
-                                  'clip_on': True},
-                    'grid': kwargs['grid']}, ax=ax)
-    else:
-        sup.mirror(spectrum_top, spectrum_bottom,
-                   {'annot_kws': {'rotation': kwargs['annotation_rotation'],
-                                  'clip_on': True},
-                    'grid': kwargs['grid']}, ax=ax)
-
-    ax.set_xlim(kwargs['mz_min'], kwargs['mz_max'])
-    ax.set_ylim(-kwargs['max_intensity'], kwargs['max_intensity'])
-
-    if not kwargs['grid']:
-        ax.spines['right'].set_visible(False)
-        ax.spines['top'].set_visible(False)
-        ax.yaxis.set_ticks_position('left')
-        ax.xaxis.set_ticks_position('bottom')
-
-    text_y = 1.2 if kwargs['cosine'] else 1.15
-    for usi, spec, loc in zip([usi1, usi2], [spectrum_top, spectrum_bottom],
-                              ['Top', 'Bottom']):
-        title = ax.text(0.5, text_y, f'{loc}: {usi}',
-                        horizontalalignment='center',
-                        verticalalignment='bottom',
-                        fontsize='x-large',
-                        fontweight='bold',
-                        transform=ax.transAxes)
-        title.set_url(f'{USI_SERVER}mirror/?usi1={usi1}&usi2={usi2}')
-        text_y -= 0.04
-        subtitle = (
-            f'Precursor $m$/$z$: '
-            f'{spec.precursor_mz:.{kwargs["annotate_precision"]}f} '
-            if spec.precursor_mz > 0 else '')
-        subtitle += f'Charge: {spec.precursor_charge}'
-        subtitle = ax.text(0.5, text_y, subtitle, horizontalalignment='center',
-                           verticalalignment='bottom', fontsize='large',
-                           transform=ax.transAxes)
-        subtitle.set_url(f'{USI_SERVER}mirror/?usi1={usi1}&usi2={usi2}')
-        text_y -= 0.06
-
-    if kwargs['cosine']:
-        subtitle_score = f'Cosine similarity = {similarity:.4f}'
-        ax.text(0.5, text_y, subtitle_score, horizontalalignment='center',
-                verticalalignment='bottom', fontsize='x-large',
-                fontweight='bold', transform=ax.transAxes)
-
-    buf = io.BytesIO()
-    plt.savefig(buf, bbox_inches='tight', format=extension)
-    buf.seek(0)
-    fig.clear()
-    plt.close(fig)
-    gc.collect()
-
-    return buf
+    try:
+        import tasks
+        result = tasks.task_generate_mirror_figure.apply_async(args=[spectrum_top, spectrum_bottom, extension, kwargs], serializer="pickle")
+        return result.get()
+    except:
+        return drawing.generate_mirror_figure(spectrum_top, spectrum_bottom, extension, **kwargs)
 
 
 def _cosine(spectrum1: sus.MsmsSpectrum, spectrum2: sus.MsmsSpectrum,
@@ -321,103 +226,13 @@ def _cosine(spectrum1: sus.MsmsSpectrum, spectrum2: sus.MsmsSpectrum,
         A tuple consisting of (i) the cosine similarity between both spectra,
         and (ii) the indexes of matching peaks in both spectra.
     """
-    spec_tup1 = SpectrumTuple(
-        spectrum1.precursor_mz, spectrum1.precursor_charge, spectrum1.mz,
-        np.copy(spectrum1.intensity) / np.linalg.norm(spectrum1.intensity))
-    spec_tup2 = SpectrumTuple(
-        spectrum2.precursor_mz, spectrum2.precursor_charge, spectrum2.mz,
-        np.copy(spectrum2.intensity) / np.linalg.norm(spectrum2.intensity))
-    return _cosine_fast(spec_tup1, spec_tup2, fragment_mz_tolerance,
+    
+    return similarity.cosine(spectrum1, spectrum2, 
+                        fragment_mz_tolerance,
                         allow_shift)
 
 
-@nb.njit
-def _cosine_fast(spec: SpectrumTuple, spec_other: SpectrumTuple,
-                 fragment_mz_tolerance: float, allow_shift: bool) \
-        -> Tuple[float, List[Tuple[int, int]]]:
-    """
-    Compute the cosine similarity between the given spectra.
 
-    Parameters
-    ----------
-    spec : SpectrumTuple
-        Numba-compatible tuple containing information from the first spectrum.
-    spec_other : SpectrumTuple
-        Numba-compatible tuple containing information from the second spectrum.
-    fragment_mz_tolerance : float
-        The fragment m/z tolerance used to match peaks in both spectra with
-        each other.
-    allow_shift : bool
-        Boolean flag indicating whether to allow peak shifts or not.
-
-    Returns
-    -------
-    Tuple[float, List[Tuple[int, int]]]
-        A tuple consisting of (i) the cosine similarity between both spectra,
-        and (ii) the indexes of matching peaks in both spectra.
-    """
-    # Find the matching peaks between both spectra, optionally allowing for
-    # shifted peaks.
-    # Candidate peak indices depend on whether we allow shifts
-    # (check all shifted peaks as well) or not.
-    # Account for unknown precursor charge (default: 1).
-    precursor_charge = max(spec.precursor_charge, 1)
-    precursor_mass_diff = ((spec.precursor_mz - spec_other.precursor_mz)
-                           * precursor_charge)
-    # Only take peak shifts into account if the mass difference is relevant.
-    num_shifts = 1
-    if allow_shift and abs(precursor_mass_diff) >= fragment_mz_tolerance:
-        num_shifts += precursor_charge
-    other_peak_index = np.zeros(num_shifts, np.uint16)
-    mass_diff = np.zeros(num_shifts, np.float32)
-    for charge in range(1, num_shifts):
-        mass_diff[charge] = precursor_mass_diff / charge
-
-    # Find the matching peaks between both spectra.
-    peak_match_scores, peak_match_idx = [], []
-    for peak_index, (peak_mz, peak_intensity) in enumerate(zip(
-            spec.mz, spec.intensity)):
-        # Advance while there is an excessive mass difference.
-        for cpi in range(num_shifts):
-            while (other_peak_index[cpi] < len(spec_other.mz) - 1 and
-                   (peak_mz - fragment_mz_tolerance >
-                    spec_other.mz[other_peak_index[cpi]] + mass_diff[cpi])):
-                other_peak_index[cpi] += 1
-        # Match the peaks within the fragment mass window if possible.
-        for cpi in range(num_shifts):
-            index = 0
-            other_peak_i = other_peak_index[cpi] + index
-            while (other_peak_i < len(spec_other.mz) and
-                   abs(peak_mz - (spec_other.mz[other_peak_i]
-                       + mass_diff[cpi])) <= fragment_mz_tolerance):
-                peak_match_scores.append(
-                    peak_intensity * spec_other.intensity[other_peak_i])
-                peak_match_idx.append((peak_index, other_peak_i))
-                index += 1
-                other_peak_i = other_peak_index[cpi] + index
-
-    score, peak_matches = 0., []
-    if len(peak_match_scores) > 0:
-        # Use the most prominent peak matches to compute the score (sort in
-        # descending order).
-        peak_match_scores_arr = np.asarray(peak_match_scores)
-        peak_match_order = np.argsort(peak_match_scores_arr)[::-1]
-        peak_match_scores_arr = peak_match_scores_arr[peak_match_order]
-        peak_match_idx_arr = np.asarray(peak_match_idx)[peak_match_order]
-        peaks_used, other_peaks_used = set(), set()
-        for peak_match_score, peak_i, other_peak_i in zip(
-                peak_match_scores_arr, peak_match_idx_arr[:, 0],
-                peak_match_idx_arr[:, 1]):
-            if (peak_i not in peaks_used
-                    and other_peak_i not in other_peaks_used):
-                score += peak_match_score
-                # Save the matched peaks.
-                peak_matches.append((peak_i, other_peak_i))
-                # Make sure these peaks are not used anymore.
-                peaks_used.add(peak_i)
-                other_peaks_used.add(other_peak_i)
-
-    return score, peak_matches
 
 
 def _prepare_spectrum(spectrum: sus.MsmsSpectrum, **kwargs: Any) \
